feat: Refactor BlockColumn component to use react-spring for animation
Refactor the BlockColumn component to incorporate the react-spring library for animating the blocks. This update adds the necessary import statement for the `useSpring` and `animated` components from `@react-spring/web`. The `useSpring` hook is used to define animation properties for the blocks, including opacity and translateY. The `animated` component is used to wrap the `Box` component to enable animation. This refactoring enhances the visual appeal of the BlockColumn component by adding smooth animations to the blocks.