Update/home page (#9)
* update: デザインを反映

* chore: 不必要なimportを削除

* feat: Add modal window for creating new habits

* feat: Update BlockColumn component to dynamically change background color and display count

Refactor the BlockColumn component to dynamically change the background color based on the count of blocks. The color is determined using a predefined array of colors and the current count. Additionally, the displayed count is calculated based on the count, with a maximum of 25 blocks. The background color is calculated using the HSL color model, with the lightness value decreasing as the count increases. This update improves the visual representation of the BlockColumn component and enhances the user experience.

* feat: Refactor BlockColumn component to dynamically change background color and display count

* feat: habitの達成回数の記録とpile

* feat: Update BlockColumn component to dynamically change background color and display count

* feat: Refactor BlockColumn component to use react-spring for animation

Refactor the BlockColumn component to incorporate the react-spring library for animating the blocks. This update adds the necessary import statement for the `useSpring` and `animated` components from `@react-spring/web`. The `useSpring` hook is used to define animation properties for the blocks, including opacity and translateY. The `animated` component is used to wrap the `Box` component to enable animation. This refactoring enhances the visual appeal of the BlockColumn component by adding smooth animations to the blocks.

---------

Co-authored-by: unknown <naoki.mizuta.work@gmail.com>
Co-authored-by: NaokiMizuta <141111910+NaokiMizuta5@users.noreply.github.com>